[aGrUM] SPDX-compliant license and refreshed base folder.
[aGrUM] refreshed base folder.

+++ acttools/ActBuilderCleaning.py
@@ -25,8 +25,12 @@
 #   ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR  #
 #   OTHER DEALINGS IN THE SOFTWARE.                                        #
 #                                                                          #
-#   See the GNU Lesser General Public License (LICENSE.LGPL) and the MIT   #
-#   licence (LICENSE.MIT) for more details.                                #
+#   See LICENCES for more details.                                         #
+#                                                                          #
+#   SPDX-FileCopyrightText: Copyright 2005-2025                            #
+#       - Pierre-Henri WUILLEMIN(_at_LIP6)                                 #
+#       - Christophe GONZALES(_at_AMU)                                     #
+#   SPDX-License-Identifier: LGPL-3.0-or-later OR MIT                      #
 #                                                                          #
 #   Contact  : info_at_agrum_dot_org                                       #
 #   homepage : http://agrum.gitlab.io                                      #
@@ -36,6 +40,8 @@
 
 import os
 import shutil
+import time
+import glob
 
 from .utils import *
 
@@ -49,13 +55,35 @@ class ActBuilderCleaning(ActBuilder):
   def check_consistency(self):
     return True
 
+  def rectouch(self, suf: str) -> int:
+    n = 0
+    now = time.time()
+    for fichier in glob.glob(os.path.join(".", "**", f"*.{suf}"), recursive=True):
+      if not self.current["dry_run"]:
+        os.utime(fichier, (now, now))
+      notif(f"  + touched : {fichier}")
+      n += 1
+    return n
+
   def build(self) -> bool:
+    ops = 0
     self.run_start()
     if os.path.isdir("build"):
-      shutil.rmtree("build")
-      self.run_done()
+      notif("Removing build")
+      if not self.current["dry_run"]:
+        shutil.rmtree("build")
+      ops += 1
+    if self.current["action"] == "purge":
+      self.run_start("purging atg files")
+      ops += self.rectouch("atg")
+      self.run_start("purging swig-generetard files")
+      ops += self.rectouch("i")
+
+    if ops > 0:
+      self.run_done(f"{ops} files/folders cleaned")
     else:
       warn("nothing to do")
       self.run_failed()
+    # find all atg files and touch them
 
     return True
